Move functions to process.js
